refactor: 이미지 로직 분리
